blog: timing charts
Not sure if this makes more sense as a blog post or in our docs
but wanted to get the content out there for review and I can
move it based on feedback.  Also, I didn't include info on how
to find baseline measures but will do that in the next reviewion
based on feedback on whether we even want this posted somewhere.

Change-Id: Ib417d14eee36659d80c5ca4eeeb4646bac6af3d0
Signed-off-by: Paul Luse <paul.e.luse@intel.com>