Update patch set 4
Patch Set 4:

In general, I think it's fine to have this as a blog post.  We might also want to put it somewhere in the developer docs so there's a more permanent place to refer to (and so it doesn't get lost when it falls off the beginning of the blog list).

Patch-set: 4