Update patch set 4
Patch Set 4:

Cool, I'll fix the acronym, formatting shit and add how to find the master baseline then resubmit.  Once approved i'll duplicate with another patch to put somewhere on the doc page.  Likely in a section on how to debug issues with patches that we need to get started on... thanks

Patch-set: 4
Label: Code-Review=0