blog: relocate test hierarchy PDF
This was originally placed in doc/, but that directory is auto-generated
by the update-doc.sh script, which deletes everything in doc.

For now, just put the PDF in the root directory and fix up the blog post
link to match.

In the long term, it would be better to move the test hierarchy
documentation into the actual SPDK repository as part of the docs.

Change-Id: I3a56602872f99669a758555601918cedb48a4426
Signed-off-by: Daniel Verkamp <daniel.verkamp@intel.com>