Remove full WebEx meeting URLs.
Passwords are embedded in the meeting URLs, which defeats
the purpose of the password.  Replace it with a generic
link to intel.webex.com where attendees can then enter
the meeting number and password.

Keep the same meeting number in this commit.  Next commit
will replace these with new meeting numbers since obviously
the current meetings are compromised.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: I1f1fbbcd7cce74b4665604d2d67fa70a35a65774