Update patch set 1
Patch Set 1: Code-Review-1

Pretty cool! If we are skipping dates though we should mention on the page - maybe at the very end that "If a meeting falls on one of the following dates it will not be held that week" or something like that.  Otherwise we say they're every other week and then the listed dates don't match up anymore.  

I think this list of holidays is fine. We don't want to be too pedantic here, we can cancel any of these at any time for whatever reason :)

Patch-set: 1
Reviewer: Paul Luse <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1