Update patch set 2
Patch Set 2: Code-Review-2

Sorry our team reported this issue but I was wrong. This section is for iSCSI initiator, but our team used this setting for SPDK iSCSI target. So, we will not have to update this section. It may be better to move this section to the bottom to avoid mislead though.

Patch-set: 2
Reviewer: Gerrit User 1012251 <1012251@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-2