Switch to Teams URLs.
Intel no longer supports Skype internally, so we are
moving to Teams instead for community calls.

Just put the URLs in the page itself rather than
the distorted images.  The links are way too long
to type manually.  If we end up with bots again like
we did with freeconferencecall, we can always enable
the Teams meeting options to force people through the
lobby before admittance.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: Idd5900b2567bdcbc0b54efb285fb0f5f8c1502a6
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk.github.io/+/8537
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Ben Walker <benjamin.walker@intel.com>
Reviewed-by: Ziye Yang <ziye.yang@intel.com>
Reviewed-by: Krzysztof Karas <krzysztof.karas@intel.com>
Reviewed-by: Karol Latecki <karol.latecki@intel.com>