update-doc: update doc_version if there is actually doc update
Update doc_version file only when generated site files
are different from the original.

The site files are generated regardless of whether
the documentation has changed or not. In addition,
the SHA of the SPDK repository changes almost
every day due to changes introduced to it, so it is
very common to push 'empty' commits that contain
only commit ID updates in the _doc_version.txt file.

To avoid this, we check if the files generated from
the SPDK repository differ from those that
are currently in the spdk.github.io repo.

Signed-off-by: Kamil Godzwon <kamilx.godzwon@intel.com>
Change-Id: I15ce16b839ee814e837d993fda4d4dc32d9210b0
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk.github.io/+/12696
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>