doc: update doc output to latest spdk-ci images
Documentation updates were pushed on any changes detected in SPDK repo,
but since last year such automation was not in place. Neither is the enviroment
that was used to generate the documentation.

It has been recently added to SPDK repo in .github/workflows/gerrit-github-pages.yml.
Yet the generated documentation is vastly different due to the above. In order
to bridge the gap, this patch was created on new spdk-ci image but with old
SPDK commit (_doc_version.txt).

Next update to latest SPDK will be done by the automation.

Doxygen version was kept the same. Most of the diff in this patch comes from
different versions of graphiviz (png files) and jekyll (header for each page
in the documentation).

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@nutanix.com>
Change-Id: I8b849fd77253196f9aea8a6749bc26d63e6c28bc
Reviewed-on: https://review.spdk.io/c/spdk/spdk.github.io/+/26809
Tested-by: Tomasz Zawadzki <tomasz@tzawadzki.com>
Reviewed-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Bypass-Merge-Requirements: Tomasz Zawadzki <tomasz@tzawadzki.com>