ci: update CI description
Document was outdated. We no longer have a "test pool"
of systems for continuous. Describe the current state
of the CI which now utilizes free-tier GitHub actions.

Remove mentions about versioning and workflows for
https//spdk.io. It is version controlled, but currently
there's no workflows to test the changes and each
change must be manually approved by maintainers.

Fixes #3756

Change-Id: Id2f528db0be6cebf6fe92419bfd1ec138e32978a
Signed-off-by: Karol Latecki <karol.latecki@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk.github.io/+/27238
Tested-by: Tomasz Zawadzki <tomasz@tzawadzki.com>
Reviewed-by: Tomasz Zawadzki <tomasz@tzawadzki.com>
Reviewed-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Reviewed-by: Michal Berger <michal.berger@nutanix.com>